app/requirements.txt
# ====================================
# AI 穿搭推薦專案 - 統整版依賴套件
# 環境: Python 3.12+ | Flask 3.1+ | Docker Ready
# 更新日期: 2025-11-28
# ====================================

# ====================================
# Web 框架核心
# ====================================
Flask==3.1.2
Werkzeug==3.1.3
Jinja2==3.1.4
itsdangerous==2.2.0
click==8.1.8
blinker==1.9.0

# ====================================
# WSGI 服務器 (生產環境)
# ====================================
gunicorn==23.0.0
gevent==24.11.1
supervisor==4.2.5

# ====================================
# 資料庫連接與 ORM
# ====================================
PyMySQL==1.1.1
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
alembic==1.14.0
cryptography==44.0.0

# ====================================
# 環境變數與配置管理
# ====================================
python-dotenv==1.0.1
python-decouple==3.8
pydantic==2.10.5
pydantic-settings==2.7.0

# ====================================
# HTTP 請求與 API
# ====================================
requests==2.32.3
urllib3==2.3.0
httpx==0.28.1
certifi==2025.11.12

# ====================================
# LangChain AI 框架
# ====================================
langchain==0.3.26
langchain-core==0.3.66
langchain-community==0.3.26
langchain-google-genai==2.0.5
langchain-groq==0.2.1
langchain-openai==0.2.14

# ====================================
# AI 和機器學習 API
# ====================================
google-generativeai==0.8.3
openai==1.58.1
groq==0.12.0

# ====================================
# 資料處理與分析
# ====================================
pandas==2.2.3
numpy==2.2.1
Pillow==11.0.0
python-dateutil==2.9.0.post0
pytz==2024.2

# ====================================
# Web 功能擴展
# ====================================
Flask-CORS==5.0.0
Flask-WTF==1.2.2
WTForms==3.2.1
Flask-Limiter==3.9.0
flask-talisman==1.1.0
flask-socketio==5.4.1

# ====================================
# 快取與效能
# ====================================
redis==5.2.1
Flask-Caching==2.3.0
orjson==3.10.13

# ====================================
# 靜態資源處理
# ====================================
Flask-Assets==2.1.0
webassets==2.0
cssmin==0.2.0
jsmin==3.0.1

# ====================================
# 測試框架
# ====================================
pytest==8.3.4
pytest-flask==1.3.0
pytest-mock==3.14.0
pytest-cov==6.0.0

# ====================================
# 程式碼品質
# ====================================
black==25.11.0
flake8==7.1.1
isort==5.13.2
mypy==1.13.0

# ====================================
# 日誌與監控
# ====================================
structlog==24.4.0
colorlog==6.9.0
sentry-sdk[flask]==2.19.2
psutil==6.1.1
Flask-HealthCheck==0.1.2

# ====================================
# 安全性
# ====================================
bcrypt==4.2.1
